chore: Remove getQueryParameter code duplication
