+++ src/lib/react-cozy-helpers/QueryParameter.js
@@ -1,0 +1,13 @@
+const arrToObj = (obj = {}, [key, val = true]) => {
+  obj[key] = val
+  return obj
+}
+
+const getQueryParameter = () =>
+  window.location.search
+    .substring(1)
+    .split('&')
+    .map(varval => varval.split('='))
+    .reduce(arrToObj, {})
+
+export default getQueryParameter
