+++ src/lib/react-cozy-helpers/index.js
@@ -5,3 +5,5 @@ import modalReducer from './ModalManager'
 export default combineReducers({ modal: modalReducer })
 
 export { ModalManager, showModal } from './ModalManager'
+
+export { default as getQueryParameter } from './QueryParameter'
